refactor: use middleware to allow placing html files in src/pages
ensure navigation during development performs a full page load when the
route does not exist, allowing vite to serve an html page.

move generated html files to their correct location after build.
